restructuring routing code
